+++ panov.ilya/P1/numOfchar.cpp
@@ -1,19 +1,19 @@
 #include "numOfchar.hpp"
 
-panov::NumOfchar::NumOfchar():
-  count(0), first(0)
+ranov::NumOfchar::NumOfchar() :
+    count(0), first(0)
 {}
 
-void panov::NumOfchar::operator()(int subsequent)
+void ranov::NumOfchar::operator()(int subsequent)
 {
-  if ((first < 0 && subsequent > 0) || (first > 0 && subsequent < 0))
-  {
-    count += 1;
-  }
-  first = subsequent;
+    if ((first < 0 && subsequent > 0) || (first > 0 && subsequent < 0))
+    {
+        count += 1;
+    }
+    first = subsequent;
 }
 
-int panov::NumOfchar::operator()() const
+int ranov::NumOfchar::operator()() const
 {
-  return count;
+    return count;
 }
